[EI-131] [api] rename: persistence test 폴더명 변경
persistent -> persistence
