+++ api/src/numerical-guidance/application/command/create-indicator-board-metadata/create-indicator-board-metadata.command.handler.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { CreateIndicatorBoardMetadataCommand } from './create-indicator-board-metadata.command';
-import { CreateIndicatorBoardMetadataPort } from '../../port/persistent/create-indicator-board-metadata.port';
+import { CreateIndicatorBoardMetadataPort } from '../../port/persistence/create-indicator-board-metadata.port';
 import { IndicatorBoardMetadata } from '../../../domain/indicator-board-metadata';
 import { Transactional } from 'typeorm-transactional';
 
@@ -19,10 +19,9 @@ export class CreateIndicatorBoardMetadataCommandHandler implements ICommandHandl
     const indicatorBoardMetaData: IndicatorBoardMetadata = IndicatorBoardMetadata.createNew(
       indicatorBoardMetaDataName,
       indicatorIds,
-      memberId,
     );
 
-    await this.createIndicatorBoardMetaDataPort.createIndicatorBoardMetaData(indicatorBoardMetaData);
+    await this.createIndicatorBoardMetaDataPort.createIndicatorBoardMetaData(indicatorBoardMetaData, memberId);
 
     return indicatorBoardMetaData;
   }
